Changed RSSI type from uint32_t to uint8_t
Because the RSSI samples can only contain a value between [0, 100],
roughly speaking, using a uint8 is better for transmission and should
not truncate the real value.

+++ interface/esb.h
@@ -44,7 +44,9 @@ typedef struct esbPacket_s {
     uint8_t data[32];
   } __attribute__((packed));
   /* Written by the radio interrupt routine */
-  uint32_t rssi;
+  /* Since the value of the RSSI sample can only be between ~[0, 100] we down cast
+   * from uint32_t to uint8_t without lose of precision */
+  uint8_t rssi;
   unsigned int crc;
 } EsbPacket;
